api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Yoshioka <sss.yoshioka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
ruff = "^0.2.0"
django-extensions = "^3.2.3"
django-bootstrap5 = "^23.4"
psycopg2 = "^2.9.9"
strawberry-graphql-django = "^0.30.0"
django-cors-headers = "^4.3.1"
pyjwt = {extras = ["crypto"], version = "^2.8.0"}
litellm = "^1.32.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
